+++ resources/views/post.blade.php
@@ -2,9 +2,9 @@
 
 @section('container')
     <article>
-        <h2>{{ $post["title"] }}</h2>
-        <h5>By : {{ $post["author"] }}</h5>
-        <p>{{ $post["body"] }}</p>
+        <h2>{{ $post->title }}</h2>
+        {{-- <h5>By : {{ $post["author"] }}</h5> --}}
+        <p>{!! $post->body !!}</p>
     </article>
 
     <a href="/blog">Back to posts</a>
