+++ routes/web.php
@@ -39,4 +39,4 @@ Route::get('/about', function () {
 Route::get('/blog', [PostController::class, 'index']);
 
 // Halaman single post
-Route::get('posts/{slug}', [PostController::class, 'show']);
+Route::get('posts/{post:slug}', [PostController::class, 'show']);
